pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Alex Aniska <alexandr.anisko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
alembic = "^1.10.4"
sqlalchemy = "1.4.41"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
loguru = "^0.7.0"
pytest = "^7.3.1"
sqlalchemy-utils = "^0.41.1"
flake8 = "^6.0.0"
black = "^23.3.0"
sqlmodel = "^0.0.8"
email-validator = "^2.0.0.post2"
yarl = "^1.9.2"
configargparse = "^1.5.3"
requests = "^2.30.0"
isort = "^5.12.0"
setuptools = "^67.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
